# openapi.yaml
openapi: 3.0.0
info:
  title: SettleIn Student Accommodation API
  description: >-
    Official OpenAPI 3.0 contract for SettleIn (Team 1) — Student Housing & Accommodation Marketplace.
    This contract specifies the interface exposed to our downstream ring partner, StudySync (Team 2),
    for student residential locations, accommodation study amenities, lease timelines, verified student
    identities, and co-living group inquiries.
  version: 1.0.0
  contact:
    name: SettleIn Engineering Team (Gift Githaka, Ian Kungu, Yahya Abdi, Tiffany Olale)
    email: support@settlein.strathmore.edu
servers:
  - url: http://localhost:5000
    description: Local development backend server
  - url: https://api.settlein.strathmore.edu
    description: Production staging API server
paths:
  /api/v1/users/{id}/residence-area:
    get:
      summary: Retrieve student residential estate and campus proximity
      description: >-
        Returns the student's residential estate name, calculated distance/commute to campus,
        and nearest campus branch. Used by StudySync to identify central meetup locations
        and recommend convenient study group venues without requiring exact GPS coordinates.
        Exact building and room numbers remain private.
      operationId: getStudentResidenceArea
      tags:
        - Students & Location
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Unique integer ID of the student user.
          schema:
            type: integer
            minimum: 1
            example: 1042
      responses:
        '200':
          description: Student residential area and proximity details retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResidenceAreaResponse'
              example:
                status_code: 200
                user_id: 1042
                estate_name: "Madaraka Estate"
                nearest_campus: "Madaraka Main Campus"
                distance_to_campus: "0.8 km (~10 mins walk)"
                cached_at: "2026-09-09T08:30:00Z"
        '401':
          description: Unauthorized — Missing or invalid authentication token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Unauthorized"
                message: "Authentication token is missing or expired."
                status_code: 401
                timestamp: "2026-09-09T09:45:00Z"
        '404':
          description: User not found or residential location not configured.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Not Found"
                message: "User with ID 1042 does not have an active residential area on file."
                status_code: 404
                timestamp: "2026-09-09T09:45:00Z"
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Internal Server Error"
                message: "An unexpected error occurred while querying student residence data."
                status_code: 500
                timestamp: "2026-09-09T09:45:00Z"
...
  /api/v1/users/{id}/lease-timeline:
    get:
      summary: Retrieve a user's lease timeline and relocation dates
      description: >-
        Returns a student user's confirmed move-in date, expected arrival time, lease start date, and lease expiration date.
        Allows StudySync to mark relocation periods as unavailable or busy on group collaboration calendars.
      operationId: getStudentLeaseTimeline
      tags:
        - Students & Location
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Unique integer ID of the student user.
          schema:
            type: integer
            minimum: 1
            example: 1042
      responses:
        '200':
          description: Student lease timeline retrieved successfully from database.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaseTimelineResponse'
              example:
                status_code: 200
                user_id: 1042
                student_id: "193923"
                booking_id: 782
                accommodation_name: "Strathmore View Executive Hostels"
                lease_status: "active"
                move_in_date: "2026-08-20T08:00:00Z"
                expected_arrival_time: "10:00 AM"
                lease_start_date: "2026-09-01"
                lease_end_date: "2026-12-15"
                relocation_window_start: "2026-08-19T00:00:00Z"
                relocation_window_end: "2026-08-22T23:59:59Z"
                is_relocating: false
                booked_at: "2026-09-22T07:30:00Z"
        '401':
          description: Unauthorized — Access restricted to verified group peers or the student.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Unauthorized"
                message: "You are not authorized to view lease schedules for this user."
                status_code: 401
                timestamp: "2026-09-09T09:45:00Z"
        '404':
          description: Student or lease booking record not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Not Found"
                message: "No active lease timeline found for user ID 1042."
                status_code: 404
                timestamp: "2026-09-09T09:45:00Z"
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Internal Server Error"
                message: "Database failure when querying lease timeline."
                status_code: 500
                timestamp: "2026-09-09T09:45:00Z"
  /api/v1/users/{id}/public-profile:
    get:
      summary: Retrieve verified student public profile and university affiliation
      description: >-
        Returns verified student public identity including display name, university affiliation,
        campus branch, degree program, and verified status. Prevents duplicate credential entry
        and verifies peer identity for study collaboration.
      operationId: getStudentPublicProfile
      tags:
        - User Identity & Profiles
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Unique integer ID of the user.
          schema:
            type: integer
            minimum: 1
            example: 1042
      responses:
        '200':
          description: Public profile retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicProfileResponse'
              example:
                status_code: 200
                user_id: 1042
                display_name: "Tiffany Akello"
                university_affiliation: "Strathmore University"
                campus_branch: "Madaraka Main Campus"
                course_of_study: "Bachelor of Business Information Technology (BBIT)"
                year_of_study: 3
                is_verified_student: true
                avatar_url: "https://images.unsplash.com/photo-1534528741775-53994a69daeb?auto=format&fit=crop&w=400&q=80"
                created_at: "2025-01-15T14:22:10Z"
        '401':
          description: Unauthorized request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Unauthorized"
                message: "Missing bearer authentication token."
                status_code: 401
                timestamp: "2026-09-09T09:45:00Z"
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Not Found"
                message: "User ID 1042 does not exist."
                status_code: 404
                timestamp: "2026-09-09T09:45:00Z"
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Internal Server Error"
                message: "Unable to fetch student profile."
                status_code: 500
                timestamp: "2026-09-09T09:45:00Z"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter the JWT bearer token received upon login.
  schemas:
    ResidenceAreaResponse:
      type: object
      required:
        - status_code
        - user_id
        - estate_name
        - nearest_campus
        - distance_to_campus
        - cached_at
      properties:
        status_code:
          type: integer
          enum: [200]
          example: 200
        user_id:
          type: integer
          description: Unique student user identifier.
          example: 1042
        estate_name:
          type: string
          description: Residential estate or neighborhood name where the student currently resides.
          example: "Madaraka Estate"
        nearest_campus:
          type: string
          description: Automatically calculated nearest university campus branch.
          example: "Madaraka Main Campus"
        distance_to_campus:
          type: string
          description: Calculated walking or transit commute distance to the nearest campus.
          example: "0.8 km (~10 mins walk)"
        cached_at:
          type: string
          format: date-time
          description: ISO 8601 timestamp representing when location data was cached.
          example: "2026-09-09T08:30:00Z"
...
    LeaseTimelineResponse:
      type: object
      required:
        - status_code
        - user_id
        - booking_id
        - lease_status
        - move_in_date
        - lease_start_date
        - lease_end_date
        - relocation_window_start
        - relocation_window_end
        - is_relocating
        - booked_at
      properties:
        status_code:
          type: integer
          enum: [200]
          example: 200
        user_id:
          type: integer
          description: Unique student user identifier.
          example: 1042
        student_id:
          type: string
          description: Official university student registration number.
          example: "193923"
        booking_id:
          type: integer
          description: Unique lease or booking identifier.
          example: 782
        accommodation_name:
          type: string
          description: Name of the booked accommodation property.
          example: "Strathmore View Executive Hostels"
        lease_status:
          type: string
          enum:
            - pending
            - confirmed
            - active
            - terminated
            - completed
          description: Current lifecycle state of the accommodation lease.
          example: "active"
        move_in_date:
          type: string
          format: date-time
          description: Exact scheduled move-in arrival timestamp (ISO 8601).
          example: "2026-08-20T08:00:00Z"
        expected_arrival_time:
          type: string
          description: Expected time of arrival on move-in day.
          example: "10:00 AM"
        lease_start_date:
          type: string
          format: date
          description: Official contractual tenancy start date.
          example: "2026-09-01"
        lease_end_date:
          type: string
          format: date
          description: Official contractual tenancy expiration date.
          example: "2026-12-15"
        relocation_window_start:
          type: string
          format: date-time
          description: Beginning of the student's moving window for calendar unavailability.
          example: "2026-08-19T00:00:00Z"
        relocation_window_end:
          type: string
          format: date-time
          description: End of the student's moving window for calendar unavailability.
          example: "2026-08-22T23:59:59Z"
        is_relocating:
          type: boolean
          description: True if the current server timestamp falls inside the relocation window.
          example: false
        booked_at:
          type: string
          format: date-time
          description: Real-time ISO 8601 timestamp when the booking was submitted and recorded in database.
          example: "2026-09-22T07:30:00Z"
    PublicProfileResponse:
      type: object
      required:
        - status_code
        - user_id
        - display_name
        - university_affiliation
        - campus_branch
        - course_of_study
        - year_of_study
        - is_verified_student
        - avatar_url
        - created_at
      properties:
        status_code:
          type: integer
          enum: [200]
          example: 200
        user_id:
          type: integer
          description: Unique student user identifier.
          example: 1042
        display_name:
          type: string
          description: Full display name of the verified student.
          example: "Tiffany Akello"
        university_affiliation:
          type: string
          description: Official university or college institution name.
          example: "Strathmore University"
        campus_branch:
          type: string
          enum:
            - "Madaraka Main Campus"
            - "Town Campus"
            - "Parklands Campus"
          description: Specific branch campus the student attends.
          example: "Madaraka Main Campus"
        course_of_study:
          type: string
          description: Academic degree program or course.
          example: "Bachelor of Business Information Technology (BBIT)"
        year_of_study:
          type: integer
          minimum: 1
          maximum: 6
          description: Current academic year of study (1 to 6).
          example: 3
        is_verified_student:
          type: boolean
          description: Flag stating whether university affiliation has been authenticated.
          example: true
        avatar_url:
          type: string
          format: uri
          description: Direct HTTPS link to student profile picture.
          example: "https://images.unsplash.com/photo-1534528741775-53994a69daeb?auto=format&fit=crop&w=400&q=80"
        created_at:
          type: string
          format: date-time
          description: Account registration timestamp.
          example: "2025-01-15T14:22:10Z"
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - status_code
        - timestamp
      properties:
        error:
          type: string
          description: Short error classification title.
          example: "Bad Request"
        message:
          type: string
          description: Human-readable explanation of why the request failed.
          example: "Field 'member_student_ids' must contain at least 2 students."
        status_code:
          type: integer
          description: HTTP status code corresponding to the error.
          example: 400
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp when the error occurred.
          example: "2026-09-09T09:45:00Z"